requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
coloredlogs==15.0.1
colorlog==6.9.0
Django==5.2.4
django-ninja==1.4.3
flatbuffers==25.2.10
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
rapidocr==3.3.1
rapidocr-onnxruntime==1.2.3
requests==2.32.4
shapely==2.1.1
six==1.17.0
some-package==0.1
sqlparse==0.5.3
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
